Disabled HE support
This driver doesnt have HE support. This is bogus behavior and a code flaw. First it said  in ECRNX_HE_CAPABILITIES in ecrnx_defs.h but then it defines  ecrnx_he_cap.has_he = true in ecrnx_he_init function in the ecrnx_main.c which is not the intended behavior and a code flaw